Refactored model classes.

+++ src/main/java/com/jayton/admissionoffice/model/to/Status.java
@@ -1,13 +1,10 @@
 package com.jayton.admissionoffice.model.to;
 
-/**
- * Created by Jayton on 30.11.2016.
- */
 public enum Status {
     CREATED,
     APPROVED;
 
-    public static Status getByOrdinal(int key) {
+    public static Status getByOrdinal(byte key) {
         switch (key) {
             case 0:
                 return CREATED;
